Show empty commit info if there is an Error
